Met en commun la gestion de bac.
